+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## [Unreleased]
+- remove heading field, provided from Core in [#53][https://github.com/DekodeInteraktiv/hogan-core/pull/53]
+
 ## 1.0.3
 - Added post type classname.
 - Added style for Savage Icons and Avatars.